+++ home.admin/99updateMenu.sh
@@ -347,10 +347,10 @@ Do you really want to update Core Lightning now?
       fi
       ;;
     RECKLESS)
-      whiptail --title "RECKLESS Core Lightning UPDATE to ${clLatestVersion}" --yes-button "Cancel" --no-button "Update" --yesno "Using the 'RECKLESS' Core Lightning update will simply
-grab the latest Core Lightning release published on the Core Lightning GitHub page (also release candidates).
+      whiptail --title "RECKLESS Core Lightning UPDATE to ${clLatestVersion}" --yes-button "Cancel" --no-button "Update" \
+      --yesno "Using the 'RECKLESS' Core Lightning update will download the latest Core Lightning release published on the Core Lightning GitHub page.
 
-There will be no security checks on signature, etc.
+The update was not tested as a part of the release.
 
 This update mode is only recommended for testing and
 development nodes with no serious funding.
@@ -367,6 +367,10 @@ Do you really want to update Core Lightning now?
         whiptail --title "ERROR" --msgbox "${error}" 8 30
       else
         echo "# Core Lightning was updated successfully"
+
+        # unlock or fix issues from the logs
+        /home/admin/config.scripts/cl.hsmtool.sh unlock ${chain}net
+
         exit 0
       fi
       ;;
@@ -553,7 +557,7 @@ case $CHOICE in
     ;;
   ELECTRS)
     /home/admin/config.scripts/bonus.electrs.sh update
-    ;;    
+    ;;
   RTL)
     /home/admin/config.scripts/bonus.rtl.sh update
     ;;